requirements.txt
alembic==1.0.6
aniso8601==4.1.0
bcrypt==3.1.6
cffi==1.11.5
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
pycparser==2.19
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
pytz==2018.9
six==1.12.0
SQLAlchemy==1.2.16
Werkzeug==0.14.1
